Added Cannonball weight/wind adjustment
In the Cannonball scene, you can adjust a slider for how much the cannonball weighs. You can also change the strength and direction of the wind by using the slider/buttons.

The cannonball also has a particle effect that makes it look like the cannonball breaks into pieces when it hits the ground.